+++ src/metadata/definitions/param-metadata.ts
@@ -1,4 +1,4 @@
-import { ValidatorOptions } from "class-validator";
+import type { ValidatorOptions } from "class-validator";
 
 import { TypeValueThunk, TypeOptions } from "../../decorators/types";
 import { ResolverData } from "../../interfaces";